requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
requests>=2.31.0
wikipedia>=1.4.0
pyperclip>=1.8.2
python-vlc>=3.0.0
pywin32>=306
streamlit>=1.28.0
pandas>=2.1.0
plotly>=5.15.0
matplotlib>=3.7.0
fpdf2>=2.7.0
beautifulsoup4>=4.12.0
